#76 Feat : 카테고리 컴포넌트 분리
